+++ src/sortLayer.ts
@@ -1,4 +1,4 @@
-import type { CanvasLayer } from './CanvasBase';
+import type { CanvasLayer } from './context';
 
 export default function sortLayers(layers: CanvasLayer[]): CanvasLayer[] {
 	return [...layers].sort(([, a], [, b]) => a - b);
